% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_calc.R
\name{dec.calc}
\alias{dec.calc}
\title{Decomposition starting from Gamma and Sigma}
\usage{
dec.calc(Gamma, Sigma)
}
\arguments{
\item{Gamma}{the \eqn{\Gamma'} matrix.}

\item{Sigma}{the \eqn{\Sigma} matrix.}
}
\value{
A list with components \itemize{
\item \code{C}: the matrix highlighting the interdependent mechanisms at deterministic level.
\item \code{Psi1}: the matrix highlighting the interdependent mechanisms at stochastic level.
\item \code{Psi0}: the matrix highlighting the causal mechanisms.
\item \code{powers}: a list containing the matrix powers of \eqn{\Gamma'}.
}
}
\description{
Function to decompose \eqn{\Gamma'} into recursive and
interdependent sub-matrices (internal use)
}
