% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroIT.R
\docType{data}
\name{macroIT}
\alias{macroIT}
\title{Italian Macroeconomic Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 11 columns.}
\source{
http://dati.istat.it/
}
\usage{
data(macroIT)
}
\description{
Italian macroeconomic variables from Q3-1996 to Q2-2011 (T = 60 observations).
The variables are
\itemize{
\item \code{QTR}: quarter and year of the observation
\item \code{C}: expenses for consumption for Italian families
\item \code{CP}: value added
\item \code{WP}: private wages from dependent employment
\item \code{I}: gross investment
\item \code{K}: gross capital stock
\item \code{GDP}: gross domestic product
\item \code{T}: taxes
\item \code{CP_1}: lagged value added
\item \code{K_1}: lagged gross capital stock
\item \code{GDP_1}: lagged gross domestic product
}
}
\examples{
data(macroIT)
}
\keyword{datasets}
