% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_calc.R
\name{rho.calc}
\alias{rho.calc}
\title{Rho Calculation}
\usage{
rho.calc(l, Gamma, A, Sigma)
}
\arguments{
\item{l}{the equation index for which to calculate rho}

\item{Gamma}{the \eqn{\Gamma'} matrix}

\item{A}{the \eqn{A} matrix}

\item{Sigma}{the \eqn{\Sigma} matrix}
}
\value{
A list with components \itemize{
\item \code{S0}: the selection matrix for \eqn{p_j}.
\item \code{S1}: the selection matrix for \eqn{\Gamma'}.
\item \code{S2}: the selection matrix. for \eqn{A}
}
}
\description{
Function to calculate rho (internal use)
}
